chore: rename expression-language folder
