+++ packages/form-js-viewer/src/features/expressionLanguage/variableExtractionHelpers.js
