+++ packages/form-js-viewer/src/features/index.js
@@ -2,6 +2,6 @@ export { default as ExpressionLanguageModule } from './expression-language';
 export { default as MarkdownModule } from './markdown';
 export { default as ViewerCommandsModule } from './viewerCommands';
 
-export * from './expression-language';
+export * from './expressionLanguage';
 export * from './markdown';
 export * from './viewerCommands';
